All tpm CI's for HMMs
The EM likelihood is the same except with no parameter transformations.
I added a likelihood function that takes all parameters on the natural
scale and use it to get the hessian.  I then use a constraint matrix to
get the correct SEs for the tpm parameters.

Next up, delta method for the stationary distributions and turning HSMM
CI's back on for the tpm.  No idea what to do about the tps that are not
directly maximized when nstates>2