Rcpp support
Rcpp support for both HMMs and HSMMs., both in fitting models and
obtaining CIs.  60% increase in speed for the last HSMM example.  Gains
probably bigger for long time series and many behavioral states.  I
could move more code to C++, but that part of the code is trickier and
I'm not sure if the gain is worth the effort.